+++ managers.tf
@@ -137,7 +137,7 @@ resource "aws_instance" "managers" {
 
   metadata_options {
     http_endpoint = "enabled"
-    http_tokens   = "required"
+    http_tokens   = var.metadata_http_tokens_required ? "required" : "optional"
   }
 
   ebs_optimized = true