+++ variables.tf
@@ -22,6 +22,11 @@ variable "create_daemon_certificate_request" {
   default     = true
 }
 
+variable "metadata_http_tokens_required" {
+  description = "Whether or not the metadata service requires session tokens, also referred to as Instance Metadata Service Version 2 (IMDSv2)."
+  default     = true
+}
+
 variable "vpc_id" {
   description = "The VPC that will contain the swarm."
 }