+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 * Allow disabling `associate_public_ip_address` to limit swarm access to other systems like API Gateway.  Enabled by default.
 * `sns_kms_id` allows encrypting the SNS messages.
+* `metadata_http_tokens_required` to allow the use of the IMDSv2 to be optional.  `true` by default.
 
 ## 6.0.1